+++ R/pdfunc.R
@@ -1,12 +1,8 @@
-#' pdfunc: compute probability density function.
+#' pdfunc: compute the probability density function.
 #'
-#' @param data A data frame containing the features (predictors).
 #' @param x A matrix of numeric features.
 #' assumed to be anomalous.
 #' @param univariate Logical indicating whether the univariate pdf should be computed.
-#' @param na.action A function specifying the action to be taken if NAs are
-#' found.
-#' @param ... Currently not used.
 #'
 #' @return A list containing the following:
 #'   \item{univariate}{TODO}
@@ -17,12 +13,21 @@
 #' @details
 #'
 #' \code{pdfunc} computes univariate or multivariate probabilities for a set of observations.
-#' @references
-#' \url{https://www.coursera.org/learn/machine-learning}
 
-#' @rdname pdfunction
+#' @rdname pdfunc
 #' @export
-pdfunc <- function() {
-  .univariate_pdf(x,x_mean,x_sd)
-  .multivariate_pdf(x,x_mean,x_sd)
+pdfunc <- function(x, univariate = TRUE) {
+
+  #check univariate
+  if (!is.logical(univariate)) stop("univariate must be logical.")
+
+  #compute mean and sd
+  x_mean <- .mean2(x)
+  x_sd <- .sd2(x)
+
+  if (univariate == TRUE) {
+    .univariate_pdf(x,x_mean,x_sd)
+  } else {
+    .multivariate_pdf(x,x_mean,x_sd)
+  }
 }